requirements.txt
pydantic==2.7.1
python-rapidjson==1.10
orjson==3.10.3
